+++ src/core/admin-file/index.ts
@@ -1,0 +1,6 @@
+export { AdminFile } from './admin-file.service';
+export { AdminFileModule } from './admin-file.module';
+export type {
+  AdminFileCreateResult,
+  FilePayload,
+} from './interfaces/file-payload.interface';