+++ src/modules/process-ifc/process-ifc.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { AuthModule } from 'src/core/config/auth.module/auth.module';
+import { ProcessIfcController } from './controllers/processifc/processifc.controller';
+import { ProcessIfcService } from './services/processifc/processifc.service';
+
+@Module({
+  imports: [AuthModule],
+  controllers: [ProcessIfcController],
+  providers: [ProcessIfcService],
+})
+export class ProcessIfcModule {}